dumping ubnt bootloader files in here

+++ bootloader/u-boot/common/cmd_sata.c
@@ -52,8 +52,8 @@ int __sata_initialize(void)
 		sata_dev_desc[i].lba = 0;
 		sata_dev_desc[i].blksz = 512;
 		sata_dev_desc[i].log2blksz = LOG2(sata_dev_desc[i].blksz);
-		sata_dev_desc[i].block_read = (int)sata_read;
-		sata_dev_desc[i].block_write = (int)sata_write;
+		sata_dev_desc[i].block_read = sata_read;
+		sata_dev_desc[i].block_write = sata_write;
 
 		rc = init_sata(i);
 		if (!rc) {
